requirements.txt
langchain
qdrant-client
fastembed
streamlit
langchain-core
langchain-openai
python-dotenv